refactor: offload WS stop, track job workers, split token format
Move WS finish_stream onto spawn_blocking+catch_unwind so Stop never
blocks the reactor or skips panic isolation. Register job-queue workers
on the shared TaskTracker so graceful drain waits for in-flight jobs.
Extract TokenFormatter/stitch and de-dupe biaser/postprocess helpers in
the engine without changing decode behavior.
