+++ crates/gigastt-core/src/inference/mod.rs
@@ -14,6 +14,7 @@ mod engine;
 mod features;
 mod pool;
 mod state;
+mod token_format;
 mod types;
 
 #[cfg(not(feature = "__internals"))]
